requirements.txt
# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# ML and Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
nltk>=3.8.1
spacy>=3.7.0
joblib>=1.3.0
shap>=0.42.0

# AI Fairness (optional)
aif360>=0.5.0
tensorflow>=2.13.0; platform_system != "Windows"  # TensorFlow not supported on Windows
tensorflow-cpu>=2.13.0; platform_system == "Windows"  # Use CPU version on Windows

# API and Web
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0 